Fallback login/signup and few fixes
