+++ app/scripts/components/account/account-manage-sub-users.components.jsx
@@ -1,5 +1,9 @@
 import React, {PropTypes} from 'react';
 import {Link} from 'react-router';
+import {graphql} from 'react-apollo';
+import gql from 'graphql-tag';
+
+import HoodieApi from '../../services/hoodie.services.js';
 
 import DisplayWithLabel from '../shared/display-with-label.components';
 
@@ -179,4 +183,62 @@ AccountManageSubUsers.propTypes = {
 	onRemoveUser: PropTypes.func,
 };
 
-export default AccountManageSubUsers;
+const withLoader = (Component) => {
+	return function Loader({loading, ...rest}) {
+		if (loading) {
+			return <p>Loading...</p>;
+		}
+
+		return <Component {...rest} />;
+	};
+};
+
+const query = gql`
+	query getSubUsers {
+		user {
+			id
+			subUsers {
+				id
+				email
+			}
+			pendingSubUsers {
+				id
+				email
+			}
+		}
+	}
+`;
+
+export default graphql(query, {
+	options: {
+		fetchPolicy: 'cache-and-network',
+	},
+	props: ({data}) => {
+		if (data.loading || !data.user) { // TMP: don't fail if there's no graphcool account
+			return {loading: true};
+		}
+
+		const members = [
+			...data.user.subUsers.map((e) => {
+				return {...e, status: 'active'};
+			}),
+			...data.user.pendingSubUsers.map((e) => {
+				return {...e, status: 'pending'};
+			}),
+		].sort((a, b) => a.email > b.email);
+
+		return {
+			members,
+			onAddUser: async (infos) => {
+				await HoodieApi.addManagedUser(data.user.id, infos);
+
+				await data.refetch();
+			},
+			onRemoveUser: async ({id}) => {
+				await HoodieApi.removeManagedUser(data.user.id, id);
+
+				await data.refetch();
+			},
+		};
+	},
+})(withLoader(AccountManageSubUsers));
